requirements.txt
asgiref==3.3.1
Django
pytz==2021.1
sqlparse==0.4.1
djangorestframework==3.13.1
django-model-utils
